HARP-14465: Fix world coordinates precision for screen space techniques (#2152)
* HARP-14465: Add IBCT test.


* HARP-14465: Store text/icon world coordinates in double precision.

POI and text positions, which use world coordinates (in contrast to tile coordinates used
by meshes), were being decoded into a single precision buffer, hence losing precision.

+++ @here/harp-mapview/lib/geometry/TileGeometryCreator.ts
@@ -490,7 +490,7 @@ export class TileGeometryCreator {
                 }
 
                 const positions = new THREE.BufferAttribute(
-                    new Float32Array(text.positions.buffer),
+                    new Float64Array(text.positions.buffer),
                     text.positions.itemCount
                 );
